+++ NEWS.md
@@ -1,3 +1,29 @@
+# ezcox 0.3.0
+
+* Remove `parallel` option in `ezcox()`
+* Add `ezcox_parallel()` for better parallel computation (#10)
+
+# ezcox 0.2.5
+
+* Add 'auto' option for filtering
+
+# ezcox 0.2.4
+
+* Fix bug: cannot get model result from parallel model
+* Add `clean_model_dir()`
+
+# ezcox 0.2.3
+
+* Add `show_models()`
+
+# ezcox 0.2.2
+
+* Add `filter_ezcox()` and `get_models()`
+
+# ezcox 0.2.1
+
+* Add `verbose` option. This may speed up parallel computation.
+
 # ezcox 0.2.0
 
 * Parallel computation is implemented in `parallel` option by **furrr** package.